Adds instructions to execute the application locally
The POM file has been fixed to allow the local execution of the
application. It was also changed to remove the "wildfly-embedded-mysql"
profile, as it is no longer useful, and to skip the test execution when
using the "wildfly-mysql-run" profile.

Additionally, the README.md file has been extended to include the
instructions related to the application execution.