Replaces SessionResource with UserResource
The SessionResource has been replaced by the UserResource. The
SessionResource's "check" method was not in accordance with the REST API
philosophy, as it represented an action instead of a resource.

As this was the only method of the SessionResource class, this class has
been removed and the UserResource class was added. Now, client's can use
this new resource to check the user's credentials through the "role"
resource, that checks the user's credentials and returns the
corresponding user role.