Fixes the CrossDomain annotation and related filters
The CrossDomain annotation included two attributes with the same
function: allowedHeaders and exposedHeaders. However, only the
"allowedHeaders" was used.

This bug caused the DefaultPatientResource to not including the
"Access-Control-Allow-Headers" header in the response, that was needed
to allow the client applications to access the "X-Total-Count" header.
In addition, the other user resources have been configured to also
expose this header.