Fixes CORS filter to correctly return requested headers
When using CORS, the client can specify some headers that should be
visible or the server application developer can specify some headers
that should be visible by default. This is controlled by the
Access-Control-Request-Headers and Access-Control-Allow-Headers.

The management of these headers was implemented in the
CrossDomainInterceptor that manages the CrossDomain annotation but it
had an error causing Access-Control-Allow-Header to be always empty.

This commit fixes this error, making the CrossDomain annotation work
properly again.