Removes games' results date limitations
The start and end date of the games's results were required to be before
the current date. This caused some problems with the differences between
client and server clocks, rejecting the results when the client's clock
is behind the server's clock.

This commit removes this limitation to avoid result rejection.